% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alleleSummary.R
\name{alleleSummary}
\alias{alleleSummary}
\title{Allele sharing summary}
\usage{
alleleSummary(x, ids)
}
\arguments{
\item{x}{An object of class \code{genomeSim}, i.e. a list of simulated
chromosomes. Each chromosome is a list, with one entry for each individual.
Each of these entries is a list of two matrices (one for each strand). The
matrices have 2 columns (start position; allele) and one row for each
segment unbroken by recombination.}

\item{ids}{A vector of ID labels. If missing, all individuals are included.}
}
\value{
A numerical matrix. Each row corresponds to a chromosomal segment.
The first 4 columns describe the segment (chromosome, start, end, length),
and are followed by two columns (paternal allele, maternal allele) for each
of the selected individuals. If \code{length(ids) == 2} two additional columns
are added:
\itemize{
\item \code{IBD} : The IBD status of each segment (= number of alleles shared
identical by descent). For a given segment, the IBD status is either 0, 1,
2 or NA. If either individual is inbred, they may be autozygous in a
segment, in which case the IBD status is reported as NA. With inbred
individuals the \code{Sigma} column (see below) is more informative than the
\code{IBD} column.
\item \code{Sigma} : The condensed identity ("Jacquard") state of each segment,
given as an integer in the range 1-9. The numbers correspond to the
standard ordering of the condensed states. In particular, for non-inbred
individuals the states 9, 8, 7 correspond to IBD status 0, 1, 2
respectively.
}
}
\description{
This function facilitates downstream analysis of simulations produced by
\code{\link[=ibdsim]{ibdsim()}}. It summarises a single genome simulation by describing the allele
flow through the pedigree.
}
\examples{
### Sibling simulation (3 sims of chromosomes 1 and 2)
x = pedtools::nuclearPed(2)
sim = ibdsim(x, sims = 3, chromosomes = 1:2)

sim1 = sim[[1]] # the first simulation

alleleSummary(sim1) # Summary of all individuals
alleleSummary(sim1, ids = 3:4) # Summary of the siblings
alleleSummary(sim1, ids = 1) # Summary of father. Trivial!

# Full sib mating: all 9 states are possible
y = pedtools::fullSibMating(1)
sim = ibdsim(y, sims = 1, chrom = 1, seed = 36)[[1]]
a = alleleSummary(sim, ids = 5:6)

stopifnot(setequal(a[, 'Sigma'], 1:9))

}
