% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_utils.R
\name{convertPos}
\alias{convertPos}
\title{Conversion of genetic map positions}
\usage{
convertPos(
  chrom = NULL,
  Mb = NULL,
  cM = NULL,
  map = "decode19",
  sex = c("average", "male", "female")
)
}
\arguments{
\item{chrom}{(Optional) A vector of chromosome labels.}

\item{Mb}{A vector of physical positions (in Mb), or NULL.}

\item{cM}{A vector of genetic positions (in cM), or NULL.}

\item{map}{A \code{genomeMap}, a \code{chromMap}, or a data frame with columns \code{Mb} and
\code{cM}. By default, \code{loadMap("decode19")} is used.}

\item{sex}{Either "average", "male" or "female".}
}
\value{
A vector of the same length as the input.
}
\description{
Convert between physical position (in megabases) and genetic position
(centiMorgan) given a chromosome map. Linear interpolation is used between
map points. Values outside the map are returned as 0, the terminal genetic
position, or NA, depending on the conversion direction.
}
\examples{
# Chromosome 1 of the built-in recombination map
map = loadMap(chrom = 1)[[1]]
head(map$male)

# Conversion Mb -> cM
phys = 1:5
gen = convertPos(Mb = phys, map = map, sex = "male")
gen

# Convert back (note that the first position cannot be recovered in this example)
convertPos(cM = gen, map = map, sex = "male")

}
