% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoLocusIdentity.R
\name{estimateIdentity}
\alias{estimateIdentity}
\alias{estimateTwoLocusIdentity}
\alias{estimateOneLocusIdentity}
\title{Estimate one-locus and two-locus identity coefficients}
\usage{
estimateTwoLocusIdentity(
  x,
  ids,
  rho = NULL,
  cM = NULL,
  Nsim,
  Xchrom = F,
  verbose = F,
  ...
)

estimateOneLocusIdentity(x, ids, Nsim, Xchrom = F, verbose = F, ...)
}
\arguments{
\item{x}{A pedigree in the form of a \code{\link[pedtools:ped]{pedtools::ped()}} object.}

\item{ids}{A character (or coercible to character) of length 2.}

\item{rho}{A scalar in the interval \verb{[0, 0.5]}: the recombination fraction
between the two loci, converted to centiMorgan using Haldanes map function:
cM = -50 * log(1 - 2*rho). Either \code{rho} or \code{cM} (but not both) must be
non-NULL.}

\item{cM}{A non-negative number: the genetic distance between the two loci,
given in centiMorgans. Either \code{rho} or \code{cM} (but not both) must be
non-NULL.}

\item{Nsim}{The number of simulations.}

\item{Xchrom}{A logical indicating if the loci are X-linked (if TRUE) or
autosomal (FALSE).}

\item{verbose}{A logical.}

\item{...}{Further arguments passed on to \code{\link[=ibdsim]{ibdsim()}}, e.g. \code{seed}.}
}
\value{
\code{estimateOneLocusIdentity()} returns a numeric vector of length 9,
with the estimated coefficients.

\code{estimateTwoLocusIBD()} returns symmetric, numerical 9*9 matrix, with the
estimated two-locus coefficients.
}
\description{
Estimate by simulation the 9 (condensed) identity coefficients of a pair
pedigree members, and also the 9*9 matrix of \emph{two-locus} identity
coefficients at a given recombination rate. (For a description of the
identity states, see here: \url{https://github.com/magnusdv/ribd}.
}
\examples{
x = fullSibMating(1)
rho = 0.25
Nsim = 100 # Increase!

estimateTwoLocusIdentity(x, ids = 5:6, rho = rho, Nsim = Nsim)

}
