% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateInbreeding.R
\name{estimateInbreeding}
\alias{estimateInbreeding}
\alias{estimateTwoLocusInbreeding}
\alias{estimateOneLocusInbreeding}
\title{Estimate one-locus and two-locus inbreeding coefficients}
\usage{
estimateTwoLocusInbreeding(
  x,
  id,
  rho = NULL,
  cM = NULL,
  Nsim,
  Xchrom = F,
  verbose = F,
  ...
)

estimateOneLocusInbreeding(x, id, Nsim, Xchrom = F, verbose = F, ...)
}
\arguments{
\item{x}{A pedigree in the form of a \code{\link[pedtools:ped]{pedtools::ped()}} object.}

\item{rho}{A scalar in the interval \verb{[0, 0.5]}: the recombination fraction
between the two loci, converted to centiMorgan using Haldanes map function:
cM = -50 * log(1 - 2*rho). Either \code{rho} or \code{cM} (but not both) must be
non-NULL.}

\item{cM}{A non-negative number: the genetic distance between the two loci,
given in centiMorgans. Either \code{rho} or \code{cM} (but not both) must be
non-NULL.}

\item{Nsim}{The number of simulations.}

\item{Xchrom}{A logical indicating if the loci are X-linked (if TRUE) or
autosomal (FALSE).}

\item{verbose}{A logical.}

\item{...}{Further arguments passed on to \code{\link[=ibdsim]{ibdsim()}}, e.g. \code{seed}.}
}
\value{
A single numeric
}
\description{
Estimate by simulation the inbreeding coefficients of any
pedigree member, and also the \emph{two-locus} inbreeding
coefficient at a given recombination rate.
}
\examples{
x = halfCousinPed(0, child = TRUE)
rho = 0.25
Nsim = 100 # Increase!

estimateTwoLocusInbreeding(x, id = 6, rho = rho, Nsim = Nsim)

}
