% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realisedCoeffs.R
\name{extractIds}
\alias{extractIds}
\title{Extract ID labels from simulation output}
\usage{
extractIds(sim)
}
\arguments{
\item{sim}{Output from \code{\link[=ibdsim]{ibdsim()}}}
}
\value{
A character vector
}
\description{
Extract ID labels from simulation output
}
\examples{
s = ibdsim(nuclearPed(2), N=1, ids = 3:4)
stopifnot(all(extractIds(s) == c("3", "4")))

}
