% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karyogram.R
\name{karyo_haploid}
\alias{karyo_haploid}
\title{Haploid karyogram}
\usage{
karyo_haploid(segments, colorBy = NA, color = "black", alpha = 1,
  bgcol = "gray99", title = NULL)
}
\arguments{
\item{segments}{A data.frame (or an object coercible to data.frame)
containing the segments to be shown on the karyogram. The first three
columns must contain chromosome (with or without "chr" prefix), start
position and stop position (in Mb). Any further columns are ignored, except
possibly a column indicated by \code{colorBy}.}

\item{colorBy}{The name of a single column of \code{segments}, to be used for
coloring. If NA (default), all segments will have the same color,
controlled by the \code{color} parameter.}

\item{color}{A single fill color for all the segments, or (if \code{colorBy}
is not NA) a named vector of colors. In the latter case, the names should
include all entries in the \code{colorBy} column.}

\item{alpha}{A single numeric in [0,1] indicating color transparency.}

\item{bgcol}{The background color of the chromosomes.}

\item{title}{Plot title.}
}
\value{
The plot object is returned invisibly, so that additional ggplot
  layers may be added if needed.
}
\description{
Show chromosomal segments in a haploid karyogram
}
\examples{

\dontrun{
segs = data.frame(chrom = c(1,4,5,5,10,10), start=c(100,50,20,80,10,50), 
                  end = c(120,100,25,100,70,120), IBD=c("cousin1","cousin2"))
karyo_haploid(segs, color="cyan")
karyo_haploid(segs, colorBy="IBD", color=c(cousin1="blue", cousin2="red"))

# To see overlaps, reduce alpha:
karyo_haploid(segs, colorBy="IBD", color=c(cousin1="blue", cousin2="red"), alpha=0.6)

# Example showing simulated IBD segments of full siblings
x = pedtools::nuclearPed(2)
s = ibdsim(x, sims=1)
a = tibble::as_tibble(alleleSummary(s[[1]], 3:4, ibd.status=TRUE))
karyo_haploid(subset(a,ibd>0), colorBy="ibd", color=c("1"="blue", "2"="red"), alpha=.3)
}

}
