% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchApp}
\alias{launchApp}
\title{Launch the ibdsim2 app}
\usage{
launchApp()
}
\value{
No return value, called for side effects.
}
\description{
This launches the Shiny app for simulating IBD segment distributions.
}
\details{
The app uses several suggested packages which are not installed by default.
To install them together with \code{ibdsim2}, use
\code{install.packages("ibdsim2", dependencies = TRUE)}.
}
\examples{

\dontrun{
launchApp()
}

}
