% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_constructors.R
\name{loadMap}
\alias{loadMap}
\title{Load a built-in genetic map}
\usage{
loadMap(map = "decode19", chrom = 1:22, uniform = FALSE, sexAverage = FALSE)
}
\arguments{
\item{map}{The name of the wanted map, possibly abbreviated. Default:
"decode19".}

\item{chrom}{A vector containing a subset of the numbers 1,2,...,23,
indicating which chromosomes to load. As a special case, \code{chrom = "X"} is
synonymous to \code{chrom = 23}. Default: \code{1:22} (the autosomes).}

\item{uniform}{A logical. If FALSE (default), the complete inhomogeneous map
is used. If TRUE, a uniform version of the same map is produced, i.e., with
the correct physical range and genetic lengths, but with constant
recombination rates along each chromosome.}

\item{sexAverage}{A logical, by default FALSE. If TRUE, a sex-averaged map is
returned, with equal recombination rates for males and females.}
}
\value{
An object of class \code{genomeMap}, which is a list of \code{chromMap}
objects. A \code{chromMap} is a list of two matrices, named "male" and "female",
with various attributes:
\itemize{
\item \code{physStart}: The first physical position (Mb) on the chromosome covered
by the map
\item \code{physEnd}: The last physical position (Mb) on the chromosome covered by
the map
\item \code{physRange}: The physical map length (Mb), equal to \code{physEnd - physStart}
\item \code{mapLen}: A vector of length 2, containing the centiMorgan lengths of the
male and female strands
\item \code{chrom}: A chromosome label
\item \code{Xchrom}: A logical. This is checked by \code{ibdsim()} and other function, to
select mode of inheritance
}
}
\description{
This function loads one of the built-in genetic maps. Currently, the only
option is the fine-scale human recombination map published by Halldorsson
et al. (2019). For reasons of speed and efficiency, the map has been
thinned down to ~14,000 data points.
}
\details{
NOTE: The built-in map was updated in version 2.3.0, adding more accurate
physical chromosome endpoints. While still based on Halldorsson et al.
(2019), the new version also uses a better thinning algorithm, allowing to
reduce the number of data points from ~38,000 to ~14,000 without losing
accuracy. The old version is available as a separate dataset
\code{legacy_decode19} for backwards reproducibility.

By setting \code{uniform = TRUE}, a uniform version of the map is returned, in
which each chromosome has the same genetic lengths as in the original, but
with constant recombination rates. This gives much faster simulations and may
be preferable in some applications.
}
\examples{
# By default, the complete map of all 22 autosomes is returned
loadMap()

# Uniform version
m = loadMap(uniform = TRUE)
m

# Check chromosome 1
m1 = m[[1]]
m1
m1$male
m1$female

# The X chromosome
loadMap(chrom = "X")[[1]]

}
\references{
Halldorsson et al. \emph{Characterizing mutagenic effects of
recombination through a sequence-level genetic map.} Science (2019).
}
\seealso{
\code{\link[=uniformMap]{uniformMap()}}, \code{\link[=customMap]{customMap()}}
}
