% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSegDist.R
\name{plotAutozygosity}
\alias{plotAutozygosity}
\title{Segment distribution plot}
\usage{
plotAutozygosity(
  ...,
  lab = NULL,
  col = NULL,
  alpha = 1,
  ellipse = TRUE,
  title = NULL,
  xlab = "Segment count",
  ylab = "Average segment length (cM)",
  legendInside = TRUE
)
}
\arguments{
\item{\dots}{One or several data frames, typically produced by
\link{realisedInbreeding}.}

\item{lab}{A character vector of labels used in the legend. Only relevant if
\code{x} is an unnamed list of length > 1.}

\item{col}{A vector of the same length as \code{...}}

\item{alpha}{A transparency parameter for the scatter points.}

\item{ellipse}{A logical: Should confidence ellipses be added to the plot?}

\item{title, xlab, ylab}{Title and axis labels. Set to \code{NULL} to remove.}

\item{legendInside}{A logical controlling the legend placement.}
}
\description{
Segment distribution plot
}
\examples{
#################################################################
# EXAMPLE
# Comparison of autozygosity distributions in various individuals
# with the same expected inbreeding coefficient (f = 1/8)
#################################################################

G = swapSex(linearPed(2), 5)           # grandfather/granddaughter
G = addChildren(G, 1, 5, 1)
HSpat = swapSex(halfSibPed(), 5)       # paternal half sibs
HSpat = addChildren(HSpat, 4, 5, 1)
HSmat = swapSex(HSpat, 1)              # maternal half sibs
QHFC = quadHalfFirstCousins()          # quad half first cousins
QHFC = addChildren(QHFC, 9, 10, nch = 1)

peds = list(G = G, HSpat = HSpat, HSmat = HSmat, QHFC = QHFC)
plotPedList(peds, newdev = TRUE)
dev.off()

# Simulations (increase N!))
s = lapply(peds, function(p)
  ibdsim(p, N = 10, ids = leaves(p), map = "uniform.sex.spec", verbose = FALSE))

# Summarise autozygous regions
segs = lapply(s, realisedInbreeding)

# Plot distributions
plotAutozygosity(segs, title = "Distribution of autozygous segments")

}
