% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSegDist.R
\name{plotSegDist}
\alias{plotSegDist}
\title{Segment distribution plot}
\usage{
plotSegDist(
  segDist,
  labels = NULL,
  alpha = 1,
  ellipses = TRUE,
  legend_inside = TRUE,
  title = NULL,
  xlab = "Segment count",
  ylab = "Average segment length (cM)"
)
}
\arguments{
\item{segDist}{A data frame or a list of data frames}

\item{labels}{A character vector, used if \code{segDist} is an unnamed list of length > 1}

\item{alpha}{A transparency parameter for the scatter points.}

\item{ellipses}{A logical: Should confidence ellipses be added to the plot?}

\item{legend_inside}{A logical indicating wether the legend should be placed
inside (default) or outside the plot window.}

\item{title, xlab, ylab}{Title and axis labels. Set to \code{NULL} to remove.}
}
\description{
Segment distribution plot
}
\examples{
#################################################################
# EXAMPLE
# Comparison of autozygosity distributions in various individuals
# with the same expected inbreeding coefficient (f = 1/8)
#################################################################

G = swapSex(linearPed(2), 5)           # grandfather/granddaughter
G = addChildren(G, 1, 5, 1)
HSpat = swapSex(halfSibPed(), 5)       # paternal half sibs
HSpat = addChildren(HSpat, 4, 5, 1)
HSmat = swapSex(HSpat, 1)              # maternal half sibs
QHFC = quadHalfFirstCousins()          # quad half first cousins
QHFC = addChildren(QHFC, 9, 10, nch = 1)

peds = list(G = G, HSpat = HSpat, HSmat = HSmat, QHFC = QHFC)
plotPedList(peds, newdev = TRUE)

# Simulations (increase 'sims'!))
s = lapply(peds, ibdsim, sims = 10, map = "uniform.sex.spec")

# Summarise autozygous regions
segs = lapply(s, realisedAutozygosity)

# Plot distributions
plotSegDist(segs, title = "Distribution of autozygous segments")

}
