% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karyogram.R
\name{prepare_segments}
\alias{prepare_segments}
\title{prepare input 'segments' for plotting (extract and rename relevant columns, merge adjacent segments)}
\usage{
prepare_segments(segments, colorBy = NA)
}
\description{
prepare input 'segments' for plotting (extract and rename relevant columns, merge adjacent segments)
}
