% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realised_ibd.R
\name{realised_kappa}
\alias{realised_kappa}
\title{Realised relatedness}
\usage{
realised_kappa(sim, ids)
}
\arguments{
\item{sim}{A list of genome simulations, as output by \code{\link[=ibdsim]{ibdsim()}}.}

\item{ids}{A vector of length 2, with ID labels of the two individuals in question.}
}
\description{
Compute the realised values of various measures of pairwise relatedness, in
simulated data. For now, only the realised IBD ("kappa") coefficients are implemented.
}
\details{
DRAFT:
Consider two members A and B of a pedigree P. The \emph{kinship coefficient}
between A and B is defined as the probability that a random allele sampled in
A is identical by descent (IBD) with an allele sampled in B at the same
autosomal locus. If this probability is taken conditional only on the
pedigree P, the result is the traditional pedigre-based kinship coefficient.

However, because of the discrete nature of meiotic recombination, the actual
IBD distribution of individuals with the specified relationship is subject to
variation. Hence we may also be interested in the \emph{realised} (or
\emph{genomic}) kinship coefficient between A and B, which is the same
probability as above, but conditional on the recombination events in the
meioses between A and B in P. For example, if the recombition events between
A and B happen to result in no segments of IBD sharing, the realised kinship
is 0, whatever the pedigree-based kinship may be.
}
\examples{
x = nuclearPed(2)
s = ibdsim(x, sims = 10)
realised_kappa(s, ids = 3:4)

}
