% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_utils.R
\name{segmentStats}
\alias{segmentStats}
\title{Summary statistics for identified segments}
\usage{
segmentStats(x, quantiles = c(0.025, 0.5, 0.975))
}
\arguments{
\item{x}{A list of matrices, whose column names must include \code{length}.}

\item{quantiles}{A vector of quantiles to include in the summary.}
}
\value{
A list of two numeric matrices, named \code{perSim} and \code{summary}. The row
names of both matrices are as follows:
\itemize{
\item \code{Count}: The total number of segments in a simulation
\item \code{Total}: The total sum of the segment lengths in a simulation
\item \code{Average}: The average segment lengths in a simulation
\item \code{Shortest}: The length of the shortest segment in a simulation
\item \code{Longest}: The length of the longest segment in a simulation
\item \code{Overall}: (Only in \code{summary}) This summarises the lengths of all
segments from all simulations
}
}
\description{
Compute summary statistics for segments identified by \code{\link[=findPattern]{findPattern()}}.
}
\examples{
x = nuclearPed(3)
sims = ibdsim(x, N = 10, map = uniformMap(M = 1), model = "haldane", seed = 1729)

# Segments where all siblings carry the same allele
segs = findPattern(sims, pattern = list(carriers = 3:5))

# Summarise
segmentStats(segs)

}
\seealso{
\code{\link[=findPattern]{findPattern()}}
}
