% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_ibd.R
\name{zero_ibd}
\alias{zero_ibd}
\title{Probability of zero ibd}
\usage{
zero_ibd(sim, id.pair, truncate = 0)
}
\arguments{
\item{sim}{A list of genome simulations, as output by \code{\link[=ibdsim]{ibdsim()}}.}

\item{id.pair}{A vector of length 2, with ID labels of the two individuals in
question.}

\item{truncate}{A numeric vector. Only IBD segments longer than this are
included in the computation. If \code{truncate} has more than one
element, a separate estimate is provided for each value. The default is to
include all segments.}
}
\value{
A data.frame with tree numeric columns:
\describe{
\item{truncate}{Same as input.}
\item{zeroprob}{The estimated probability of no IBD segments.}
\item{SE}{The standard error of the estimate.}
}
}
\description{
Estimate the probability of no ibd sharing in a pairwise relationship.
}
\examples{
### 
# The following example computes the probability of
# no IBD sharing between a pair of fourth cousins.
# We also show how the probability is affected by 
# truncation, i.e., ignoring short segments.
###

# Define the pedigree
x = pedtools::cousinsPed(4)
cousins = pedtools::leaves(x)

# Simulate (increase 'sims'!)
s = ibdsim(x, sims = 7)

# Probability of zero ibd segments. (By default all segs are used)
zero_ibd(s, id.pair = cousins)

# Re-compute the probability with several truncation levels
truncate = 0:50
zp = zero_ibd(s, id.pair = cousins, truncate = truncate)

plot(truncate, zp$zeroprob)

}
