% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alleleSummaryX.R
\name{alleleSummaryX}
\alias{alleleSummaryX}
\title{X-chromosomal allele sharing summary}
\usage{
alleleSummaryX(x, ids)
}
\arguments{
\item{x}{An object of class \code{genomeSim}, i.e. a list of simulated
chromosomes. Each chromosome is a list, with one entry for each individual.
Each of these entries is a list of two matrices (one for each strand). The
matrices have 2 columns (start position; allele) and one row for each
segment unbroken by recombination.}

\item{ids}{A vector of ID labels. If missing, all individuals are included.}
}
\value{
A numerical matrix. Each row corresponds to a chromosomal segment.
The first 4 columns describe the segment (chromosome, start, end, length).,
and are followed by one or two columns for each of the selected
individuals: One column (maternal allele) for males and two columns
(paternal allele, maternal allele) for females. If \code{length(ids) == 2} two
additional columns are added:
\itemize{
\item \code{IBD} : The IBD status of each segment (= number of alleles shared
identical by descent). For a given segment, the IBD status is either 0, 1,
2 or NA. If either individual is X-inbred, they may be autozygous in a
segment, in which case the IBD status is reported as NA. With inbred
individuals the \code{Sigma} column (see below) is more informative than the
\code{IBD} column.
\item \code{Sigma} : The condensed identity state of each segment, given as an
integer in the range 1-9. The numbers refer to the \emph{autosomal} states in
the usual ordering; for details about this, and how how they relate to
identity states on X, please see the explanation at the \code{ribd} homepage:
\url{https://github.com/magnusdv/ribd}.
}
}
\description{
This is the X-chromosomal version of \code{\link[=alleleSummary]{alleleSummary()}}.
}
\examples{
x = fullSibMating(1)
s = ibdsim(x, N = 1, ids = NULL, chrom = 23)[[1]]

# Complete summary.
# Note only one allele column for males
alleleSummaryX(s)

# Outbred brother/sister
alleleSummaryX(s, ids = 3:4)

# Inbred brother/sister
alleleSummaryX(s, ids = 5:6)

}
