% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawHaplo.R
\name{haploDraw}
\alias{haploDraw}
\title{Draw haplotypes onto a pedigree plot}
\usage{
haploDraw(
  x,
  ibd,
  pos,
  cols = NULL,
  height = NA,
  width = NA,
  sep = 0.75 * width,
  dist = 1.5 * width,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{A \code{ped} object.}

\item{ibd}{A \code{genomeSim} object.}

\item{pos}{A vector of the same length as \code{labels(x)}, indicating where
haplotypes should be drawn relative to the pedigree symbols: 0 = no
haplotypes; 1 = below; 2 = left; 3 = above; 4 = right.}

\item{cols}{A colour vector corresponding to the alleles in \code{ibd}.}

\item{height, width}{Positive numbers; the height/width of a haplotype
rectangle}

\item{sep}{The separation between haplotypes within a pair.}

\item{dist}{The distance between pedigree symbols and the closest haplotype.}

\item{verbose}{A logical. If TRUE, plotting parameters are printed to the
screen. This may be useful for reproducing plots made interactively.}

\item{...}{Arguments passed on to \code{plot.ped()}.}
}
\value{
None.
}
\description{
Visualise the IBD pattern of a single chromosome, by drawing haplotypes onto
the pedigree.
}
\examples{
op = par()

###############################
# Example 1: A family quartet #
###############################

x = nuclearPed(2)
s = ibdsim(x, N = 1, map = uniformMap(M = 1), seed = 4276)
s[[1]]

haploDraw(x, s[[1]], pos = c(2,4,1,1), cols = c(3,7,2,4),
          margin = c(3, 5, 3, 5), cex = 1.2)


###########################
# Example 2: Autozygosity #
###########################

x = halfCousinPed(0, child = TRUE)
s = ibdsim(x, N = 1, map = uniformMap(M = 1),
           skipRecomb = spouses(x, 1), seed = 19499)
s[[1]]

# Gray colour (8) for irrelevant founder alleles
haploDraw(x, s[[1]], pos = c(1,0,2,0,4,4),
          cols = c(2,4,8,8,8,8), margin = c(2, 2, 2, 2))


# Restore graphics parameters
par(op)

}
