% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibdsim.R
\name{ibdsim}
\alias{ibdsim}
\title{IBD simulation}
\usage{
ibdsim(x, sims, query = NULL, condition = NULL, map = "decode",
  chromosomes = NULL, model = "chi", merged = TRUE, simdata = NULL,
  skip.recomb = "noninf_founders", seed = NULL, verbose = TRUE)
}
\arguments{
\item{x}{A pedigree in the form of a \code{\link{linkdat}} object.}

\item{sims}{A positive integer indicating the number of simulations.}

\item{query, condition}{A single allele pattern (SAP), i.e., a list with
numerical entries named "0", "1", "2", "atleast1", "atmost1".}

\item{map}{The genetic map(s) to be used in the simulations: One of the
character strings "decode", "uniform.sex.spec", "uniform.sex.aver". (See
Details.)}

\item{chromosomes}{A numeric vector indicating chromosome numbers, or either
of the words "AUTOSOMAL" or "X". The default is 1:22, i.e. the human
autosomes.}

\item{model}{A character indicating the statistical model for recombination:
Either "chi" (the default) or "haldane". (See details.)}

\item{merged}{A logical, indicating if overlapping/adjacent regions should be
merged or not.}

\item{simdata}{Either NULL, in which case simulation is performed before
collecting IBD regions, or an object containing simulation data (typically
generated by a previous run of \code{IBDsim}).}

\item{skip.recomb}{A numeric containing individuals whose meioses should be
simulated without recombination (i.e. a random strand is passed on to each
offspring). If NULL, nobody is skipped. The default value (the character
"noninf_founders") computes the set of pedigree founders that cannot be
carriers of the alleles described in the \code{condition} and/or
\code{query} SAPs.}

\item{seed}{An integer to be passed on to \code{\link{set.seed}}).}

\item{verbose}{A logical.}
}
\value{
If the \code{query} is NULL, no identification of IBD segments is
  done, and only the simulated genomes are (invisibly) returned. If
  \code{query} is non-null, the segments with the query SAP are identified,
  and a list of three elements is returned. These are the 'simdata' (the
  simulated genomes), the 'segments' (a list of matrices describing all
  identified regions) and finally 'stats' (a matrix with one column per
  simulation, summarizing the regions from that genome). A summary is printed
  on the screen, with some or all of the following slots:

  \describe{
  \item{count.all}{The average count of all IBD segments (i.e. counting both
  random regions and the disease region in case of conditional simulation).}
  \item{fraction.all}{The average fraction (in \%) of the genome covered by
  IBD segments.}
  \item{average.all}{The average length (in Mb) of IBD segments.}
  \item{longest.all}{The average length (in Mb) of the longest IBD segment
  from each simulated genome.}
  \item{count.rand}{The average count of random IBD segments.}
  \item{fraction.rand}{The average fraction (in \%) of the chromosomes
  covered by random IBD segments.}
  \item{average.rand}{The average length (in Mb) of random IBD segments.}
  \item{longest.rand}{The average length (in Mb) of the longest random IBD
  segment from each simulated genome.}
  \item{length.dis}{The average length of the disease segment (only with
  conditional simulation).}
  \item{rank.dis}{The average rank of the disease segment length among all
  the segments.}
  \item{zeroprob}{The percentage of simulations resulting in zero IBD
  segments.} 
  }

  The term 'IBD segment' in the above always refers to 'IBD segment matching
  the query SAP'. The suffixes \code{.dis}, \code{.rand} and \code{.all}
  points to respectively the \emph{disease} IBD segments (only relevant in
  conditional simulations), the \emph{random}, i.e. non-disease, IBD segments
  (only relevant in conditional simulations), and \emph{all} IBD segments (in
  any simulation).
}
\description{
This is the main function of the package. Gene dropping and recombination of
chromosomes is simulated down the pedigree, either unconditionally or
conditional on given allele patterns. Regions satisfying the 'query' pattern
are collected and summarized.
}
\details{
Each simulation starts by unique alleles being distributed to the pedigree
founders. In each subsequent meiosis, homologue chromosomes are made to
recombine according to a renewal process along the four-strand bundle, with
chi square distributed waiting times. (For comparison purposes, Haldane's
Poisson model for recombination is also implemented.)

Recombination rates are sex-dependent, and vary along each chromosome
according to the recombination map specified by the \code{map} parameter. By
default, the complete Decode map of the human autosome is used (see
References). If \code{map="uniform.sex.spec"}, the genetic chromosome
\emph{lengths} are as in the Decode map, but the recombination rate is kept
constant along each chromosome. If \code{map="uniform.sex.aver"}, sex
averaged genetic chromosome lengths are used (and constant recombination
rates along each chromosome).

IBD patterns are described as combinations of Single Allele Patterns (SAPs).
A SAP is a specification for a given allele of the number of copies carried
by various individuals, and must be given as a list of numerical vectors
(containing ID labels) named '0', '1', '2', 'atleast1' and 'atmost1' (some of
these can be absent or NULL; see Examples).

If a condition SAP is given (i.e. if \code{condition} is non-null),
simulation of each complete chromosome set (all autosomes by default) is
performed as follows: A 'disease chromosome' is sampled at random (using the
physical chromosome lengths as weights), followed by a random 'disease locus'
on this chromosome.  For this chromosome, gene dropping down the pedigree is
carried out in such a way that the 'disease locus' has the condition SAP. (In
a bit more detail: First, the program computes all possible sets of obligate
carriers, with suitable weights, and samples one of these.  Included in the
obligate carriers will e exactly one founder, one of whose alleles is taken
as the 'disease allele'. In each meiosis involving obligate carriers,
recombination is performed as usual, but the strand carrying the 'disease
allele' is always the one passed on.) For the other chromosomes, simulation
is done unconditionally.
}
\examples{
# Half first cousins.
# Two such cousins will on average share 1/8 = 12.5\\\% of the autosome.

z = paramlink::halfCousinPed(1)
plot(z)
res = ibdsim(z, sims=1, query=list('1'=8:9))
res$stats
# visualizing the spread in total IBD sharing and the number of segments:

IBD_percent = res$stats['fraction.all', ]
IBD_count = res$stats['count.all', ]
hist(IBD_percent)
hist(IBD_count)

}
