% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibdsim.R
\name{ibdsim}
\alias{ibdsim}
\title{IBD simulation}
\usage{
ibdsim(
  x,
  sims = 1,
  map = "decode",
  chromosomes = NULL,
  model = "chi",
  skipRecomb = NULL,
  seed = NULL,
  verbose = TRUE
)
}
\arguments{
\item{x}{A \code{\link[pedtools:ped]{pedtools::ped()}} object.}

\item{sims}{A positive integer indicating the number of simulations.}

\item{map}{The genetic map to be used in the simulations: One of the
character strings "decode", "uniform.sex.spec", "uniform.sex.aver". (See
Details.)}

\item{chromosomes}{A numeric vector indicating chromosome numbers, or either
of the words "AUTOSOMAL" or "X". The default is 1:22, i.e., the human
autosomes.}

\item{model}{Either "chi" (default) or "haldane", indicating the statistical
model for recombination. (See details.)}

\item{skipRecomb}{An optional vector of ID labels indicating individuals
whose meioses should be simulated without recombination. (Each child will
then receive a random strand of each chromosome.) If NULL (default), nobody
is skipped. Note that \code{skipRecomb = founders(x)} is a sensible (and
time saving) choice in many applications.}

\item{seed}{An integer to be passed on to \code{\link[=set.seed]{set.seed()}}).}

\item{verbose}{A logical.}
}
\value{
The simulated genomes are invisibly returned.
}
\description{
This is the main function of the package, simulating the recombination
process in each meioses of a pedigree. The output summarises the IBD segments
between all or a subset of individuals.
}
\details{
Each simulation starts by unique alleles being distributed to the pedigree
founders. In each meiosis, homologue chromosomes are made to recombine
according to a renewal process along the four-strand bundle, with chi square
distributed waiting times. (For comparison purposes, Haldane's Poisson model
for recombination is also implemented.)

Recombination rates are sex-dependent, and vary along each chromosome
according to the recombination map specified by the \code{map} parameter. By
default, the complete Decode map of the human autosome is used (see
References). If \code{map = "uniform.sex.spec"}, the genetic chromosome \emph{lengths}
are as in the Decode map, but the recombination rate is kept constant along
each chromosome. If \code{map = "uniform.sex.aver"}, sex averaged genetic
chromosome lengths are used (and constant recombination rates along each
chromosome).
}
\examples{

hs = halfSibPed()
ibdsim(hs, sims = 2, map = uniformMap(M = 1))


}
