% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibdsim.R
\name{ibdsim}
\alias{ibdsim}
\title{IBD simulation}
\usage{
ibdsim(
  x,
  sims,
  condition = NULL,
  map = "decode",
  chromosomes = NULL,
  model = "chi",
  skip.recomb = "noninf_founders",
  seed = NULL,
  verbose = TRUE
)
}
\arguments{
\item{x}{A pedigree in the form of a \code{\link[pedtools:ped]{pedtools::ped()}} object.}

\item{sims}{A positive integer indicating the number of simulations.}

\item{condition}{A single allele pattern (SAP), i.e., a list with
numerical entries named "0", "1", "2", "atleast1", "atmost1".}

\item{map}{The genetic map(s) to be used in the simulations: One of the
character strings "decode", "uniform.sex.spec", "uniform.sex.aver". (See
Details.)}

\item{chromosomes}{A numeric vector indicating chromosome numbers, or either
of the words "AUTOSOMAL" or "X". The default is 1:22, i.e. the human
autosomes.}

\item{model}{A character indicating the statistical model for recombination:
Either "chi" (the default) or "haldane". (See details.)}

\item{skip.recomb}{A numeric containing individuals whose meioses should be
simulated without recombination (i.e. a random strand is passed on to each
offspring). If NULL, nobody is skipped. The default value (the character
"noninf_founders") computes the set of pedigree founders that cannot be
carriers of the alleles described in the \code{condition} SAPs.}

\item{seed}{An integer to be passed on to \code{\link[=set.seed]{set.seed()}}).}

\item{verbose}{A logical.}
}
\value{
The simulated genomes are invisibly returned.
}
\description{
This is the main function of the package. Gene dropping and recombination of
chromosomes is simulated down the pedigree, either unconditionally or
conditional on given allele patterns.
}
\details{
Each simulation starts by unique alleles being distributed to the pedigree
founders. In each subsequent meiosis, homologue chromosomes are made to
recombine according to a renewal process along the four-strand bundle, with
chi square distributed waiting times. (For comparison purposes, Haldane's
Poisson model for recombination is also implemented.)

Recombination rates are sex-dependent, and vary along each chromosome
according to the recombination map specified by the \code{map} parameter. By
default, the complete Decode map of the human autosome is used (see
References). If \code{map="uniform.sex.spec"}, the genetic chromosome
\emph{lengths} are as in the Decode map, but the recombination rate is kept
constant along each chromosome. If \code{map="uniform.sex.aver"}, sex
averaged genetic chromosome lengths are used (and constant recombination
rates along each chromosome).

IBD patterns are described as combinations of Single Allele Patterns (SAPs).
A SAP is a specification for a given allele of the number of copies carried
by various individuals, and must be given as a list of vectors containing ID
labels, named '0', '1', '2', 'atleast1' and 'atmost1' (some of
these can be absent or NULL; see Examples).

If a condition SAP is given (i.e. if \code{condition} is non-null),
simulation of each complete chromosome set (all autosomes by default) is
performed as follows: A 'disease chromosome' is sampled at random (using the
physical chromosome lengths as weights), followed by a random 'disease locus'
on this chromosome.  For this chromosome, gene dropping down the pedigree is
carried out in such a way that the 'disease locus' has the condition SAP. (In
a bit more detail: First, the program computes all possible sets of obligate
carriers, with suitable weights, and samples one of these.  Included in the
obligate carriers will e exactly one founder, one of whose alleles is taken
as the 'disease allele'. In each meiosis involving obligate carriers,
recombination is performed as usual, but the strand carrying the 'disease
allele' is always the one passed on.) For the other chromosomes, simulation
is done unconditionally.
}
\examples{

hs = halfSibPed()
res = ibdsim(hs, sims = 2, map = uniformMap(M = 1))
res
alleleSummary(res[[1]])

}
