% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibdsim2-package.R
\docType{package}
\name{ibdsim2}
\alias{ibdsim2}
\alias{ibdsim2-package}
\title{ibdsim2: Simulation of chromosomal regions shared by family members}
\description{
Simulation of segments shared identical-by-descent (IBD) by
pedigree members. Using sex specific recombination rates along the human
genome (Kong et. al (2010) \url{doi:10.1038/nature09525}), phased chromosomes
are simulated for all pedigree members, either by unconditional gene
dropping or conditional on a specified IBD pattern. Additional functions
provide summaries and further analysis of the simulated genomes.
}
