% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karyogram.R
\name{karyoHaploid}
\alias{karyoHaploid}
\title{Haploid karyogram}
\usage{
karyoHaploid(
  segments,
  chrom = 1:22,
  colorBy = NA,
  color = NULL,
  separate = TRUE,
  alpha = 1,
  bgcol = "gray92",
  title = NULL,
  legendTitle = colorBy,
  prefix = "chr"
)
}
\arguments{
\item{segments}{A data.frame (or an object coercible to data.frame)
containing the segments to be shown on the karyogram. The first three
columns must contain chromosome (with or without "chr" prefix), start
position and stop position (in Mb). Any further columns are ignored, except
possibly a column indicated by \code{colorBy}.}

\item{colorBy}{The name of a single column of \code{segments}, to be used for
colouring. If NA (default), all segments will have the same colour,
controlled by the \code{color} parameter.}

\item{color}{A single fill colour for all the segments, or (if \code{colorBy} is
not NA) a named vector of colours. In the latter case, the names should
include all entries in the \code{colorBy} column.}

\item{separate}{A logical; relevant only if the \code{colorBy} column has more
than one level. If FALSE, all segments are drawn in full height. This may
not be optimal if segments of different colours overlap. If TRUE the levels
are drawn in separate bands on the chromosomes.}

\item{alpha}{A single numeric in \verb{[0,1]} indicating colour transparency.}

\item{bgcol}{The background colour of the chromosomes.}

\item{title}{Plot title.}
}
\value{
The plot object is returned invisibly, so that additional \code{ggplot}
layers may be added if needed.
}
\description{
Show chromosomal segments in a haploid karyogram
}
\examples{

\dontrun{
segs = data.frame(chrom = c(1,4,5,5,10,10),
                  start = c(100,50,20,80,10,50),
                  end = c(120,100,25,100,70,120),
                  IBD = c("cousin1","cousin2"))
cols = c(cousin1 = "blue", cousin2 = "red")

karyoHaploid(segs, color = "cyan")
karyoHaploid(segs, colorBy = "IBD", color = cols)

# Note difference if `separate = FALSE`
karyoHaploid(segs, colorBy = "IBD", color = cols, separate = FALSE)

# To see the overlap now, you can reduce alpha:
karyoHaploid(segs, colorBy = "IBD", color = cols, separate = FALSE, alpha = 0.7)
              
# Example showing simulated IBD segments of full siblings
s = ibdsim(nuclearPed(2), N = 1, ids = 3:4)
karyogram2(s[[1]])
}

}
