% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_utils.R
\name{loadMap}
\alias{loadMap}
\title{Load a built-in genetic map}
\usage{
loadMap(map = "decode19", chrom = 1:22, detailed = TRUE, sexSpecific = TRUE)
}
\arguments{
\item{map}{The name of the wanted map. Currently, the only valid choice is
"decode19". This is also the default.}

\item{chrom}{A numeric vector indicating which chromosomes to load. Default:
\code{1:22} (the autosomes)}

\item{detailed}{A logical. If TRUE (default), the complete inhomogeneous map
is used. If FALSE, a uniform version of the same map is produced, i.e. with
the correct lengths, but constant recombination rate along each chromosome.}

\item{sexSpecific}{A logical, by default TRUE. If FALSE, a sex-averaged map
is returned, equal between males and females}
}
\value{
An object of class \code{genomeMap}.
}
\description{
This function loads one of the built-in genetic maps. A faster, uniform
version is also available by the parameter \code{detailed = FALSE}.
}
\examples{
# By default, the complete map of all 22 autosomes is returned
loadMap()

# Uniform version
m = loadMap(detailed = FALSE)

# Check chromosome 1:
m1 = m[[1]]
m1$male
m1$female

}
\references{
Halldorsson et al. \emph{Characterizing mutagenic effects of recombination through a
sequence-level genetic map.} Science 363, no. 6425 (2019).
}
