% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ibd1.R
\name{plot_ibd1}
\alias{plot_ibd1}
\title{Scatter plot of IBD=1 segment distributions}
\usage{
plot_ibd1(..., labels, alpha = 1, ellipses = TRUE,
  legend_inside = TRUE)
}
\arguments{
\item{...}{One or several objects of class \code{genomeSimList}, i.e. outputs
of \code{\link[=ibdsim]{ibdsim()}}.}

\item{labels}{A character vector of the same length as above, containing
descriptive plotting labels for the inputs.}

\item{alpha}{A transparency parameter for the scatter points.}

\item{ellipses}{A logical: Should confidence ellipses be added to the plot?}

\item{legend_inside}{A logical indicating wether the legend should be placed inside (default) or outside the plot window.}
}
\value{
A ggplot2 plot object.
}
\description{
Visualise and compare distributions of segments with IBD=1 between two
individuals.
}
\details{
This function takes as input one or several complete outputs from the
\code{\link[=ibdsim]{ibdsim()}}. The underlying pedigree of each input is extracted, and
\code{\link[pedtools:leaves]{pedtools::leaves()}} is used to identify the two individuals to be
examined. (Hence it is required that each pedigree has exactly two leaves.)

For each simulation the number of IBD=1 segments is plotted against the
average segment length. Finally, contour curves are added to plot,
corresponding to the theoretical (i.e., pedigree-based) value of
\eqn{\kappa_1} for each input relationship.
}
\examples{
###
# Compare the distribution of IBD=1 segments
# between paternal and maternal half siblings.
###
library(pedtools)
# Define the pedigrees
x.pat = halfSibPed()
x.mat = swapSex(x.pat, 1)

# Simulate (increase 'sims'!)
map = "uniform.sex.spec"
sims = 10
s.pat = ibdsim(x.pat, sims = sims, map=map)
s.mat = ibdsim(x.mat, sims = sims, map=map)

plot_ibd1(s.pat, s.mat, labels = c("HSpat", "HSmat"))

}
