% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realisedCoeffs.R
\name{realised}
\alias{realised}
\alias{realisedKappa}
\alias{realisedAutozygosity}
\title{Realised relatedness}
\usage{
realisedKappa(sims, ids = attr(sims, "ids"))

realisedAutozygosity(sims, id = attr(sims, "ids"))
}
\arguments{
\item{sims}{A list of genome simulations, as output by \code{\link[=ibdsim]{ibdsim()}}.}

\item{ids}{A vector with two ID labels.}

\item{id}{A single ID label.}
}
\description{
Compute the realised values of various pedigree coefficients, in simulated
data. The current implementation covers realised inbreeding coefficients
(autozygosity) and realised kappa coefficients.
}
\details{
DRAFT: Consider two members A and B of a pedigree P. The \emph{kinship
coefficient} between A and B is defined as the probability that a random
allele sampled in A is identical by descent (IBD) with an allele sampled in B
at the same autosomal locus. If this probability is taken conditional only on
the pedigree P, the result is the traditional pedigree-based kinship
coefficient.

However, because of the discrete nature of meiotic recombination, the actual
IBD distribution of individuals with the specified relationship is subject to
variation. Hence we may also be interested in the \emph{realised} (or \emph{genomic})
kinship coefficient between A and B, which is the same probability as above,
but conditional on the recombination events in the meioses between A and B in
P. For example, if the recombination events between A and B happen to result
in no segments of IBD sharing, the realised kinship is 0, whatever the
pedigree-based kinship may be.
}
\examples{

# Realised IBD coefficients between full siblings
x = nuclearPed(2)
s = ibdsim(x, N = 10) # increase N
realisedKappa(s, ids = 3:4)

# Realised inbreeding coefficients, child of first cousins
x = cousinPed(1, child = TRUE)
s = ibdsim(x, N = 10)
realisedAutozygosity(s, id = 9)


}
