% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_utils.R
\name{uniformMap}
\alias{uniformMap}
\title{Uniform recombination maps}
\usage{
uniformMap(Mb = NULL, cM = NULL, M = NULL, cm.per.mb = 1, chromosome = 1)
}
\arguments{
\item{Mb}{Map length in megabases.}

\item{cM}{Map length in centiMorgan}

\item{M}{Map length in Morgan}

\item{cm.per.mb}{A positive number; the cM/Mb ratio}

\item{chromosome}{A chromosome label.}
}
\value{
An object of class "chromosomeMap", which is a list of two matrices,
named "male" and "female".
}
\description{
Create a uniform recombination map of a given length.
}
\examples{
uniformMap(M = 1)

}
