% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_ibd.R
\name{zeroIBD}
\alias{zeroIBD}
\title{Probability of zero IBD}
\usage{
zeroIBD(sims, ids, truncate = 0)
}
\arguments{
\item{sims}{A list of genome simulations, as output by \code{\link[=ibdsim]{ibdsim()}}.}

\item{ids}{A vector with two ID labels.}

\item{truncate}{A vector of positive real numbers. Only IBD segments longer than this are
included in the computation. If \code{truncate} has more than one
element, a separate estimate is provided for each value. The default (\code{truncate = 0}) is to
include all segments.}
}
\value{
A data frame with tree numeric columns:
\describe{
\item{truncate}{Same as input.}
\item{zeroprob}{The estimated probability of no IBD segments.}
\item{SE}{The standard error of the estimate.}
}
}
\description{
Estimate the probability of no IBD sharing in a pairwise relationship.
}
\examples{
### 
# The following example computes the probability of
# no IBD sharing between a pair of fourth cousins.
# We also show how the probability is affected by 
# truncation, i.e., ignoring short segments.
###

# Define the pedigree
x = cousinPed(4)
cousins = leaves(x)

# Simulate (increase `sims`!)
s = ibdsim(x, sims = 10)

# Probability of zero ibd segments. (By default all segs are used)
zeroIBD(s, ids = cousins)

# Re-compute the probability with several truncation levels
truncate = 0:20
zp = zeroIBD(s, ids = cousins, truncate = truncate)

plot(truncate, zp$zeroprob, type = "b", ylim = c(0,1))

}
